backend/blog/serializers/category_serializer.py
from rest_framework import serializers
from blog.models import Category


class CategorySerializer(serializers.ModelSerializer):
    posts = serializers.PrimaryKeyRelatedField(many=True, read_only=True)

    class Meta:
        model = Category
        fields = ["id", "name", "owner", "posts"]
