backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==15.3.4
mysqlclient==1.4.2.post1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
